clean up comments and dependencies
